requirements.txt
faiss-cpu==1.13.0
fastapi==0.122.0
google-genai==1.52.0
neo4j==5.28.2
numpy==2.2.5
openai==1.101.0
pdfplumber==0.11.8
pydantic==2.11.7
pydantic_core==2.33.2
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
uvicorn==0.38.0